Update pipeline to run on Docker runners
The runners connected to this GitLab instance are now Docker runners
instead of shell runners.  This means an image should be specified to
run the commands in.

The CI/CD process for this project only involves moving the files into a
public/ directory so they can be picked up by GitLab pages, for that
reason only the Alpine image is needed.

Closes #1.